+++ caddyfile_test.go
@@ -13,7 +13,6 @@ func TestDirectives(t *testing.T) {
 		valid     bool
 		backend   string
 		acl       []string
-		badger    map[string]string
 		ristretto map[string]string
 	}{
 		{
@@ -21,23 +20,6 @@ func TestDirectives(t *testing.T) {
 			valid:     true,
 			backend:   "ristretto",
 			acl:       []string{"127.0.0.1"},
-			badger:    nil,
-			ristretto: nil,
-		},
-		{
-			caddyfile: `mwcache map`,
-			valid:     true,
-			backend:   "map",
-			acl:       []string{"127.0.0.1"},
-			badger:    nil,
-			ristretto: nil,
-		},
-		{
-			caddyfile: `mwcache badger`,
-			valid:     true,
-			backend:   "badger",
-			acl:       []string{"127.0.0.1"},
-			badger:    nil,
 			ristretto: nil,
 		},
 		{
@@ -45,11 +27,9 @@ func TestDirectives(t *testing.T) {
 			valid:     false,
 			backend:   "",
 			acl:       nil,
-			badger:    nil,
 			ristretto: nil,
 		},
 		{
-			// 4
 			caddyfile: `
 			mwcache {
 				purge_acl 11.11.11.11
@@ -58,21 +38,6 @@ func TestDirectives(t *testing.T) {
 			valid:     true,
 			backend:   "ristretto",
 			acl:       []string{"11.11.11.11"},
-			badger:    nil,
-			ristretto: nil,
-		},
-		{
-			// 5
-			caddyfile: `
-			mwcache {
-				map
-				purge_acl 11.11.11.11
-			}
-			`,
-			valid:     true,
-			backend:   "map",
-			acl:       []string{"11.11.11.11"},
-			badger:    nil,
 			ristretto: nil,
 		},
 		{
@@ -87,7 +52,6 @@ func TestDirectives(t *testing.T) {
 			valid:     true,
 			backend:   "ristretto",
 			acl:       []string{"11.11.11.11", "11.11.11.12"},
-			badger:    nil,
 			ristretto: nil,
 		},
 		{
@@ -104,7 +68,6 @@ func TestDirectives(t *testing.T) {
 			valid:     true,
 			backend:   "ristretto",
 			acl:       []string{"11.11.11.11", "11.11.11.12", "11.11.11.13", "11.11.11.14"},
-			badger:    nil,
 			ristretto: nil,
 		},
 		// TODO
@@ -120,37 +83,6 @@ func TestDirectives(t *testing.T) {
 		{
 			caddyfile: `
 			mwcache {
-				badger {
-					in_memory true
-				}
-			}
-			`,
-			valid:     true,
-			backend:   "badger",
-			acl:       []string{"127.0.0.1"},
-			badger:    map[string]string{"in_memory": "true"},
-			ristretto: nil,
-		},
-		{
-			// 9
-			caddyfile: `
-			mwcache {
-				badger {
-					in_memory true
-					value_log_file_size 8388608 # 1<23
-				}
-			}
-			`,
-			valid:     true,
-			backend:   "badger",
-			acl:       []string{"127.0.0.1"},
-			badger:    map[string]string{"in_memory": "true", "value_log_file_size": "8388608"},
-			ristretto: nil,
-		},
-		{
-			//
-			caddyfile: `
-			mwcache {
 				ristretto {
 					num_counters 100000
 					max_cost 10000
@@ -161,7 +93,6 @@ func TestDirectives(t *testing.T) {
 			valid:     true,
 			backend:   "ristretto",
 			acl:       []string{"127.0.0.1"},
-			badger:    nil,
 			ristretto: map[string]string{"num_counters": "100000", "max_cost": "10000", "buffer_items": "64"},
 		},
 	}
@@ -189,13 +120,6 @@ func TestDirectives(t *testing.T) {
 			e := strings.Join(test.acl, ", ")
 			a := strings.Join(m.config.PurgeAcl, ", ")
 			t.Errorf("Test %d: Expected: '%s' but got '%s'", i, e, a)
-		}
-
-		for k, a := range m.config.BadgerConfig {
-			e := test.badger[k]
-			if a != e {
-				t.Errorf("Test %d: Expected: '%s' but got '%s'", i, e, a)
-			}
 		}
 
 		for k, a := range m.config.RistrettoConfig {